[DISCO-3017] Move top level config modules to configs folder
