+++ merino/configs/app_configs/config_sentry.py
@@ -8,7 +8,7 @@ from sentry_sdk.integrations.fastapi import FastApiIntegration
 from sentry_sdk.integrations.starlette import StarletteIntegration
 from sentry_sdk.types import Event, Hint
 
-from merino.config import settings
+from merino.configs.config import settings
 from merino.utils.version import fetch_app_version_from_file
 
 logger = logging.getLogger(__name__)
