+++ merino/main.py
@@ -10,8 +10,8 @@ from fastapi.exceptions import RequestValidationError
 from fastapi.middleware.cors import CORSMiddleware
 
 from merino import curated_recommendations, providers
-from merino.config_logging import configure_logging
-from merino.config_sentry import configure_sentry
+from merino.configs.app_configs.config_logging import configure_logging
+from merino.configs.app_configs.config_sentry import configure_sentry
 from merino.utils.metrics import configure_metrics, get_metrics_client
 from merino.middleware import featureflags, geolocation, logging, metrics, user_agent
 from merino.web import api_v1, dockerflow
