+++ merino/providers/top_picks/backends/top_picks.py
@@ -8,7 +8,7 @@ from enum import Enum
 from json import JSONDecodeError
 from typing import Any
 
-from merino.config import settings
+from merino.configs.config import settings
 from merino.exceptions import BackendError
 from merino.providers.top_picks.backends.filemanager import (
     DomainDataSource,
