+++ merino/providers/top_picks/provider.py
@@ -11,7 +11,7 @@ from collections import defaultdict
 from typing import Any
 
 from merino.utils import cron
-from merino.config import settings
+from merino.configs.config import settings
 from merino.exceptions import BackendError
 from merino.providers.base import BaseProvider, BaseSuggestion, SuggestionRequest
 from merino.providers.top_picks.backends.filemanager import GetFileResultCode
