+++ merino/utils/synced_gcs_blob.py
@@ -14,7 +14,7 @@ from google.cloud.storage import Client
 from aiodogstatsd import Client as StatsdClient
 
 from merino.utils import cron
-from merino.config import settings
+from merino.configs.config import settings
 
 logger = logging.getLogger(__name__)
