+++ tests/integration/api/v1/curated_recommendations/fakespot_backend/test_fakespot_backend.py
@@ -10,7 +10,7 @@ from aiodogstatsd import Client as StatsdClient
 from google.cloud.storage import Client, Bucket
 import orjson
 
-from merino.config import settings
+from merino.configs.config import settings
 from merino.curated_recommendations.corpus_backends.protocol import ScheduledSurfaceId
 from merino.curated_recommendations.fakespot_backend.fakespot_backend import GcsFakespot
 from merino.utils.synced_gcs_blob import SyncedGcsBlob
