+++ tests/integration/api/v1/curated_recommendations/test_curated_recommendations.py
@@ -14,7 +14,7 @@ from pydantic import HttpUrl
 from pytest_mock import MockerFixture
 from scipy.stats import linregress
 
-from merino.config import settings
+from merino.configs.config import settings
 from merino.curated_recommendations import (
     CorpusApiBackend,
     CuratedRecommendationsProvider,
