+++ tests/integration/api/v1/fake_providers.py
@@ -9,7 +9,7 @@ from typing import Protocol
 
 from pydantic import HttpUrl
 
-from merino.config import settings
+from merino.configs.config import settings
 from merino.providers import BaseProvider
 from merino.providers.base import BaseSuggestion, SuggestionRequest
